pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "reg-agent"
version = "0.1.0"
description = "HEMA tournament registration enrichment agent"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.83.0",
    "pandas>=3.0",
    "pydantic>=2.12",
    "pydantic-ai>=1.62",
    "gspread>=6.0",
    "discord>=2.3.2",
    "python-dotenv>=1.0",
    "langfuse>=3.0",
    "table2ascii>=1.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes (undefined names, unused imports, etc.)
    "B",   # flake8-bugbear (common bugs)
    "UP",  # pyupgrade (modern Python syntax)
]
ignore = [
    "E501",  # line too long (handled by formatter if needed)
    "B008",  # do not perform function calls in default args (pydantic-ai patterns use this)
    "E402",  # module import not at top (load_dotenv() must run before other imports)
]

[project.scripts]
reg-agent = "reg_agent.main:main"
